Updated Navbar, About Section and Scroll Offset
Navbar has been updated to show the Logo to the left and the links to the right.

About Section has been updated, aligning the About Alive header with the text box. The paragraph text has been changed to align to the left. Fix the Learn about our services button to be under the text and centered. The button also offsets on the scroll to the services section.